+++ simple_manga_downloader/modules/manganelo_com.py
@@ -11,7 +11,11 @@ class Manganelo(BaseManga):
     session.headers.update({"Referer": base_link})
     site_re = re.compile(r"https?://manganelo\.com/manga/\S*")
 
-    def __init__(self, link):
+    def __init__(self, link, title=None):
+        if title:
+            self.series_title = title
+        else:
+            self.series_title = None
         self.manga_link = link
         self.cover_url = None
         self.chapters = {}
@@ -31,8 +35,9 @@ class Manganelo(BaseManga):
         except AttributeError:
             return "Failed to get page"
 
-        title = soup.find(class_="story-info-right").find("h1").text
-        self.series_title = self.clean_up_string(title)
+        if self.series_title is None:
+            title = soup.find(class_="story-info-right").find("h1").text
+            self.series_title = self.clean_up_string(title)
         if title_return:
             return True